% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{Hourly series of Soil Moisture (SM) and Precipitation and (Pobs) for 16 sites.}
\format{
A list with 16 data frames, each with hourly Soil Moisture (SM) and Precipitation and (Pobs).
Each item in the list contains the following:
\describe{
  \item{SM}{Soil moisture time series, in mm}
  \item{Pobs}{Precipitation, in mm}
  ...
}
}
\source{
Personal communication with Dr. Luca Brocca (\url{http://hydrology.irpi.cnr.it/people/luca-brocca/})
}
\usage{
data
}
\description{
Dataset containing hourly series of Soil Moisture (SM) and Precipitation and (Pobs) for 16 sites (i.e., gauging stations) around the world.
}
\keyword{datasets}
