% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm2rain.R
\name{sm2rain}
\alias{sm2rain}
\title{Simulation of SM2RAIN algorithm.}
\usage{
sm2rain(PAR, data, NN = 24, deltaSM = 0.005)
}
\arguments{
\item{PAR}{A 3-dimensional vector of parameters.}

\item{data}{A dataframe with dimensions n x 3. The 1st column of the datafram should contain dates, the 2nd the Soil moisture series, and the 3rd the precipitation series.}

\item{NN}{A scalar indicating the aggregation level. Default: 24 (assuming that the the data are hourly, and we wish to aggregate them at daily temporal scale).}

\item{deltaSM}{SM2RAIN algorithm parameter. Default: 0.005.}
}
\value{
A list with 4 elements. That is, Date, SM (Soil moisture), Psim (Simulated precipitations), Pobs (Observed precipitation) and R (Corr[Psim,Pobs])
}
\description{
Simulation of SM2RAIN algorithm.
}
\examples{
# Define some parameters for the SM2RAIN algorithm
PAR = c(64.65, 2.09, 10.41)

# Use the data contained in the first element of the list called data.
ts_df = data[[1]]

# Estimate percipitation using SM2RAIN algorithm
Sim = sm2rain(PAR=PAR, data=ts_df, NN=24)
plot(Sim$Pobs, t='l', col='black')
lines(Sim$Psim, col='red' )

}
