% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm2rainNSE.R
\name{sm2rainNSE}
\alias{sm2rainNSE}
\title{Objective function for the calibration of SM2RAIN algorithm. Maxmimization of Nash-Sutcliffe Efficiency (NSE).}
\usage{
sm2rainNSE(PAR, data, NN = 24, deltaSM = 0.005)
}
\arguments{
\item{PAR}{A 3-dimensional vector of parameters.}

\item{data}{A dataframe with dimensions n x 3. The 1st column of the datafram should contain dates, the 2nd the Soil moisture series, and the 3rd the precipitation series.}

\item{NN}{A scalar indicating the aggregation level. Default: 24 (assuming that the the data are hourly, and we wish to aggregate them at daily temporal scale).}

\item{deltaSM}{SM2RAIN algorithm parameter. Default: 0.005.}
}
\value{
A scalar indicating the Nash-Sutcliffe Efficiency (NSE) of the selected parameters for the provided data.
}
\description{
Objective function for the calibration of SM2RAIN algorithm. Maxmimization of Nash-Sutcliffe Efficiency (NSE).
}
\examples{
# Define some parameters for the SM2RAIN algorithm
PAR = c(64.65, 2.09, 10.41)

# Use the data contained in the first element of the list called data.
ts_df = data[[1]]

# Estimate the NSE value comparing precipitation obtained by the
SM2RAIN algorithm (and parameters defined in PAR) and observed precipitation.

NSE = sm2rainNSE(PAR=PAR, data=ts_df, NN=24)
print(NSE)

}
